+++ packages/react-pdf-toolkit/src/components/atoms/Divider/types.ts
@@ -1,0 +1,3 @@
+import { BaseComponentProps } from '@/types/base-component-props';
+
+export interface DividerProps extends BaseComponentProps {}