+++ sdks/workflow-browser-sdk/.eslintrc.cjs
@@ -3,5 +3,8 @@ module.exports = {
   env: {
     browser: true,
   },
+  parserOptions: {
+    project: './tsconfig.eslint.json',
+  },
   extends: ['@ballerine/eslint-config'],
 };