+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/headless-example
 
+## 0.1.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.41
+  - @ballerine/workflow-browser-sdk@0.5.41
+
+## 0.1.40
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.40
+  - @ballerine/workflow-browser-sdk@0.5.40
+
 ## 0.1.39
 
 ### Patch Changes