+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/eslint-config
 
+## 1.0.5
+
+### Patch Changes
+
+- version bump
+
+## 1.0.4
+
+### Patch Changes
+
+- updated eslint rules and plugins
+
 ## 1.0.3
 
 ### Patch Changes