+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/ui
 
+## 0.3.28
+
+### Patch Changes
+
+- updating versions
+- Updated dependencies
+  - @ballerine/common@0.7.41
+
+## 0.3.27
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.40
+
 ## 0.3.26
 
 ### Patch Changes